backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
pandas
openpyxl
google-genai
pillow